Se estandariza proyecto.
