+++ src/components/common/List/index.js
@@ -1,5 +1,5 @@
-import ExampleList from './exampleList'
-import GenericList from './genericList'
-import SimpleList from './simpleList'
+import ExampleList from './ExampleList'
+import GenericList from './GenericList'
+import SimpleList from './SimpleList'
 
 export { ExampleList, GenericList, SimpleList }
